backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
Pillow==10.0.0
django-cors-headers==4.3.1
drf-spectacular==0.27.1
python-dotenv==1.0.1 